requirements.txt
appdirs==1.4.3
celery==3.1.23
dj-database-url==0.3.0
Django==1.11.1
djangorestframework==3.8.2
django-google-charts==0.1.1
django-redis==4.6.0
django-qsstats-magic==0.7.2
django-widget-tweaks==1.4.1
Fabric3==1.12.post1
gunicorn==19.9.0
Jinja2==2.9.6
kombu==3.0.37
packaging==16.8
paramiko==2.4.1
psycopg2==2.7.1
pyparsing==2.2.0
python-dateutil==2.5.3
pytz==2017.2
redis==2.10.5
requests==2.18.2
six==1.10.0
# We use aiohttp gunicorn workers
asyncio==3.4.3
aiohttp==1.3.5
aiohttp-wsgi==0.6.6
Pillow==3.4.2

# For automation check of certificate expiration
pyOpenSSL==17.2.0
